fix: putting a logic together using composable
